moved TokenType enum to separate file
