+++ src/panel/tag_parser/TokenType.js
@@ -1,0 +1,8 @@
+const TokenType = {
+	OPEN_PAREN: "OPEN_PAREN",
+	CLOSE_PAREN: "CLOSE_PAREN",
+	BINARY_OP: "BINARY_OP",
+	LITERAL: "LITERAL"
+};
+
+export default TokenType;
